fix: refactor runners to new folder structure specified on #151
