+++ .gitignore
@@ -6,4 +6,4 @@ helm/kre/values.yaml
 helm/values-dell.yaml
 helm/kre/charts/*
 operator/helm-charts/kre-chart/charts
-runtime-runners/kre-py/test/env_vars.sh
+runners/kre-py/test/env_vars.sh
