+++ runners/kre-entrypoint/.gitlab-ci.yml
@@ -1,34 +1,34 @@
-release-kre-go:
+release-kre-entrypoint:
   stage: semantic-release
   image: node:10
   before_script:
-    - npm install --prefix runtime-runners/kre-go
+    - npm install --prefix runners/kre-entrypoint
   script:
-    - cd runtime-runners/kre-go && npx semantic-release -e semantic-release-monorepo
+    - cd runners/kre-entrypoint && npx semantic-release -e semantic-release-monorepo
   rules:
     - if: '$CI_COMMIT_BRANCH == "master"'
       when: always
       changes:
-        - runtime-runners/kre-go/**/*
-        - .gitlab-ci.yml
+        - runners/kre-entrypoint/**/*
+      - .gitlab-ci.yml
 
-publish-kre-go-image:
+publish-kre-entrypoint-image:
   stage: publish-image
   image: docker:18-git
   services:
     - docker:18-dind
   variables:
-    IMAGE_PROJECT_NAME: "konstellation/kre-go"
+    IMAGE_PROJECT_NAME: "konstellation/kre-runtime-entrypoint"
     DOCKER_USER: $DOCKER_USER
     DOCKER_PASS: $DOCKER_PASS
   before_script:
     - docker login -u $DOCKER_USER -p $DOCKER_PASS
   script:
-    - TAG=${CI_COMMIT_TAG#kre-go-v}
-    - docker build -t $IMAGE_PROJECT_NAME:$TAG runtime-runners/kre-go
+    - TAG=${CI_COMMIT_TAG#kre-entrypoint-v}
+    - docker build -t $IMAGE_PROJECT_NAME:$TAG runners/kre-entrypoint
     - docker tag $IMAGE_PROJECT_NAME:$TAG $IMAGE_PROJECT_NAME:latest
     - docker push $IMAGE_PROJECT_NAME:$TAG
     - docker push $IMAGE_PROJECT_NAME:latest
   rules:
-    - if: '$CI_COMMIT_TAG =~ /^kre-go-v\d*.\d*.\d*$/'
+    - if: '$CI_COMMIT_TAG =~ /^kre-entrypoint-v\d*.\d*.\d*$/'
       when: always
