+++ runners/kre-entrypoint/entrypoint-gen/entrypoint.py.tmpl
