+++ runners/kre-entrypoint/entrypoint-gen/generate_entrypoint.go
