+++ runners/kre-entrypoint/entrypoint-gen/generate_entrypoint_test.go
