+++ runners/kre-entrypoint/entrypoint-gen/main.go
