+++ runners/kre-entrypoint/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "runtime-entrypoint",
+  "name": "kre-entrypoint",
   "repository": {
     "type": "git",
     "url": "https://github.com/konstellation-io/kre.git"
