+++ runners/kre-entrypoint/src/kre_grpc.py
