+++ runners/kre-go/.gitlab-ci.yml
@@ -1,34 +1,34 @@
-release-runtime-entrypoint:
+release-kre-go:
   stage: semantic-release
   image: node:10
   before_script:
-    - npm install --prefix runtime-entrypoint
+    - npm install --prefix runners/kre-go
   script:
-    - cd runtime-entrypoint && npx semantic-release -e semantic-release-monorepo
+    - cd runners/kre-go && npx semantic-release -e semantic-release-monorepo
   rules:
     - if: '$CI_COMMIT_BRANCH == "master"'
       when: always
       changes:
-      - runtime-entrypoint/**/*
-      - .gitlab-ci.yml
+        - runners/kre-go/**/*
+        - .gitlab-ci.yml
 
-publish-runtime-entrypoint-image:
+publish-kre-go-image:
   stage: publish-image
   image: docker:18-git
   services:
     - docker:18-dind
   variables:
-    IMAGE_PROJECT_NAME: "konstellation/kre-runtime-entrypoint"
+    IMAGE_PROJECT_NAME: "konstellation/kre-go"
     DOCKER_USER: $DOCKER_USER
     DOCKER_PASS: $DOCKER_PASS
   before_script:
     - docker login -u $DOCKER_USER -p $DOCKER_PASS
   script:
-    - TAG=${CI_COMMIT_TAG#runtime-entrypoint-v}
-    - docker build -t $IMAGE_PROJECT_NAME:$TAG runtime-entrypoint
+    - TAG=${CI_COMMIT_TAG#kre-go-v}
+    - docker build -t $IMAGE_PROJECT_NAME:$TAG runners/kre-go
     - docker tag $IMAGE_PROJECT_NAME:$TAG $IMAGE_PROJECT_NAME:latest
     - docker push $IMAGE_PROJECT_NAME:$TAG
     - docker push $IMAGE_PROJECT_NAME:latest
   rules:
-    - if: '$CI_COMMIT_TAG =~ /^runtime-entrypoint-v\d*.\d*.\d*$/'
+    - if: '$CI_COMMIT_TAG =~ /^kre-go-v\d*.\d*.\d*$/'
       when: always
