+++ runners/kre-go/.golangci.yml
