+++ runners/kre-go/context_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/nats-io/nats.go"
 	"go.mongodb.org/mongo-driver/bson"
 
-	"github.com/konstellation-io/kre/runtime-runners/kre-go/config"
-	"github.com/konstellation-io/kre/runtime-runners/kre-go/mocks"
+	"github.com/konstellation-io/kre/runners/kre-go/config"
+	"github.com/konstellation-io/kre/runners/kre-go/mocks"
 )
 
 type TestPrediction struct {
