+++ runners/kre-go/entrypoint.sh
