+++ runners/kre-py/test/env_vars.sh.tpl
@@ -5,5 +5,5 @@ export KRT_NATS_SERVER=localhost:4222
 export KRT_NATS_INPUT=test-subject-input
 export KRT_NATS_OUTPUT=
 export KRT_NATS_MONGO_WRITER=mongo_writer
-export KRT_BASE_PATH=${PWD}/runtime-runners/kre-py/test/myvol
+export KRT_BASE_PATH=${PWD}/runners/kre-py/test/myvol
 export KRT_HANDLER_PATH=src/node/node_handler.py
